# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Qu\u1ea3n l\u00fd Tao API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/widget-token": {
      "post": {
        "tags": [
          "Widget"
        ],
        "summary": "Generate Widget Token",
        "operationId": "generate_widget_token_v1_widget_token_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WidgetTokenResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Widget"
        ],
        "summary": "Delete Widget Token",
        "operationId": "delete_widget_token_v1_widget_token_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/widget/summary": {
      "get": {
        "tags": [
          "Widget"
        ],
        "summary": "Get Widget Summary",
        "operationId": "get_widget_summary_v1_widget_summary_get",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WidgetSummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WidgetSummaryResponse": {
        "properties": {
          "pending_count": {
            "type": "integer",
            "title": "Pending Count"
          },
          "pending_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Pending Ids"
          },
          "captured_epoch": {
            "type": "integer",
            "title": "Captured Epoch"
          }
        },
        "type": "object",
        "required": [
          "pending_count",
          "pending_ids",
          "captured_epoch"
        ],
        "title": "WidgetSummaryResponse"
      },
      "WidgetTokenResponse": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "expires_at": {
            "type": "string",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "token",
          "expires_at"
        ],
        "title": "WidgetTokenResponse"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}